+++ src/actions/index.js
@@ -1,0 +1,2 @@
+export const REMOVE_BOOK = book => ({ type: 'REMOVE_BOOK', book });
+export const CREATE_BOOK = book => ({ type: 'CREATE_BOOK', book });
