+++ src/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import './App.css';
-import BooksList from './containers/BooksList';
-import BooksForm from './components/BooksForm';
+import '../App.css';
+import BooksList from '../containers/BooksList';
+import BooksForm from './BooksForm';
 
 const App = () => (
   <div className="App">
