+++ src/containers/BooksList.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
-import BookList from '../components/BookList';
-import { CREATE_BOOK, REMOVE_BOOK } from '../reducers/index';
+import BookList from './BookList';
+import { CREATE_BOOK, REMOVE_BOOK } from '../actions/index';
 
 const mapStateToProps = state => ({ books: state.books });
 const mapDispatchToProps = dispatch => ({
